factoriser les écrans
